+++ X68000 Shared/px68k/win32api/peace.c
@@ -114,7 +114,7 @@ GetTickCount(void)
 	struct timeval tv;
 
 	gettimeofday(&tv, 0);
-	return tv.tv_usec / 1000 + tv.tv_sec * 1000;
+	return (DWORD)(tv.tv_usec / 1000 + tv.tv_sec * 1000);
 }
 
 BOOL WINAPI
